refactor: reduce AppShell coupling + split test_fastapi_service (#137, #139)
#137: AppShell 373→197 lines. Extract SuggestContext, useRouteSelection,
useSessionHydration, useConversationSync.

#139: test_fastapi_service.py 597→3 files + shared conftest.
